+++ Anarchy/REST/Guild/Properties/CreationProperties.cs
@@ -31,7 +31,7 @@ namespace Discord
         public Image Icon
         {
             get { return _image.Image; }
-            set { _image.Image = value; }
+            set { _image.SetImage(value); ; }
         }
         #endregion
     }