+++ Anarchy/REST/Guild/Properties/GuildProperties.cs
@@ -71,7 +71,7 @@ namespace Discord
             get { return _image.Image; }
             set
             { 
-                _image.Image = value;
+                _image.SetImage(value);
                 _useBase64 = true;
                 IconSet = true;
             }