+++ Anarchy/REST/User/Settings/UserProfile.cs
@@ -72,7 +72,7 @@ namespace Discord
             get { return _image.Image; }
             set
             {
-                _image.Image = value;
+                _image.SetImage(value);
 
                 AvatarSet = true;
             }